requirements.txt
alembic==1.0.11
attrs==19.1.0
flake8==3.7.8
flask==1.1.1
freezegun==0.3.12
inject==3.5.3
mypy==0.720
pylint==2.3.1
pytest==5.0.1
sqlalchemy==1.3.6
